+++ BALSAMIC/utils/exc.py
@@ -1,7 +1,7 @@
 class BalsamicError(Exception):
     """Base exception for the BALSAMIC."""
 
-    def __init__(self, message):
+    def __init__(self, message: str = ""):
         super().__init__(message)