+++ BALSAMIC/utils/rule.py
@@ -1,5 +1,6 @@
 import os
 import re
+import toml
 import logging
 from pathlib import Path
 import snakemake
@@ -350,3 +351,82 @@ def get_pon_samples(fastq_dir):
         (f.split("_1"))[0] for f in os.listdir(fastq_dir) if f.endswith("_R_1.fastq.gz")
     ]
     return samples
+
+
+def get_clinical_snv_observations(config: dict) -> str:
+    """Returns path for clinical snv observations
+
+    Args:
+        config: a config dictionary
+
+    Returns:
+        Path for clinical_snv_observations vcf file
+
+    """
+    return Path(config["reference"]["clinical_snv_observations"]).as_posix()
+
+
+def get_swegen_snv(config: dict) -> str:
+    """Returns path for swegen snv frequencies
+
+    Args:
+        config: a config dictionary
+
+    Returns:
+        Path for swegen_snv vcf file
+
+    """
+    return Path(config["reference"]["swegen_snv_frequency"]).as_posix()
+
+
+def get_clinical_sv_observations(config: dict) -> str:
+    """Returns path for clinical sv observations
+
+    Args:
+        config: a config dictionary
+
+    Returns:
+        Path for clinical_sv_observations vcf file
+
+    """
+    return Path(config["reference"]["clinical_sv_observations"]).as_posix()
+
+
+def get_swegen_sv(config: dict) -> str:
+    """Returns path for swegen sv frequencies
+
+    Args:
+        config: a config dictionary
+
+    Returns:
+        Path for swegen_sv vcf file
+
+    """
+    return Path(config["reference"]["swegen_sv_frequency"]).as_posix()
+
+
+def get_toml(annotation: dict) -> str:
+    """Returns annotation in toml format
+
+    Args:
+        annotation: a dict containing annotation resource
+
+    Returns:
+        toml_annotation: a string in toml format
+    """
+    return toml.dumps(annotation)
+
+
+def dump_toml(annotations: list) -> str:
+    """Returns list of converted annotation in toml format
+
+    Args:
+        annotations: a list of toml
+
+    Returns:
+        toml_annotation: a string of toml annotation resources
+    """
+    toml_annotations = ""
+    for annotation in annotations:
+        toml_annotations += get_toml(annotation)
+    return toml_annotations