+++ CHANGELOG.rst
@@ -1,3 +1,62 @@
+[11.0.0]
+-------
+
+Added:
+^^^^^^
+* Use of PON reference, if exists for CNVkit tumor-normal analysis https://github.com/Clinical-Genomics/BALSAMIC/pull/982
+* Added PON version to CLI and config.json https://github.com/Clinical-Genomics/BALSAMIC/pull/983
+* `cnvpytor` to varcallpy3 container https://github.com/Clinical-Genomics/BALSAMIC/pull/991
+* `cnvpytor` for tumor only workflow https://github.com/Clinical-Genomics/BALSAMIC/pull/994
+* R packages to cnvkit container https://github.com/Clinical-Genomics/BALSAMIC/pull/996
+* Missing R packages to cnvkit container https://github.com/Clinical-Genomics/BALSAMIC/pull/997
+* add rlang to cnvkit container https://github.com/Clinical-Genomics/BALSAMIC/pull/998
+* AnnotSV and bedtools to annotate container https://github.com/Clinical-Genomics/BALSAMIC/pull/1005
+* cosmicdb to TNscope for tumor only and tumor normal workflows https://github.com/Clinical-Genomics/BALSAMIC/pull/1006
+* `loqusDB` dump files to the config through the balsamic config case CLI https://github.com/Clinical-Genomics/BALSAMIC/pull/992
+* Pre-annotation quality filters for SNVs annd added `research` to output files https://github.com/Clinical-Genomics/BALSAMIC/pull/1007
+* Annotation of snv_clinical_observations for somatic snv https://github.com/Clinical-Genomics/BALSAMIC/pull/1012
+* Annotation of sv_clinical_observations  for somatic sv and SV CNV filter rules https://github.com/Clinical-Genomics/BALSAMIC/pull/1013
+* Swegen SNV and SV frequency database for WGS https://github.com/Clinical-Genomics/BALSAMIC/pull/1014
+* triallelic_sites and variants with MaxDepth to the VCFs https://github.com/Clinical-Genomics/BALSAMIC/pull/1021
+* Clinical VCF for TGA workflow https://github.com/Clinical-Genomics/BALSAMIC/pull/1024
+* CNVpytor plots into the CNV PDF report https://github.com/Clinical-Genomics/BALSAMIC/pull/1023
+* Research and clinical housekeeper tags https://github.com/Clinical-Genomics/BALSAMIC/pull/1023
+* Cluster configuration for rules https://github.com/Clinical-Genomics/BALSAMIC/pull/1028
+* Variant filteration using loqusDB and Swegen annotations https://github.com/Clinical-Genomics/BALSAMIC/pull/1029
+* Annotation resources to readsthedocs https://github.com/Clinical-Genomics/BALSAMIC/pull/1031
+* Delly CNV rules for TGA workflow https://github.com/Clinical-Genomics/BALSAMIC/pull/103
+* cnvpytor container and removed cnvpytor from varcallpy3 https://github.com/Clinical-Genomics/BALSAMIC/pull/1037
+
+Changed:
+^^^^^^^^
+* Added version number to the PON reference filename (`.cnn`) https://github.com/Clinical-Genomics/BALSAMIC/pull/982
+* Update `TIDDIT` to v3.3.0, `SVDB` to v2.6.4, `delly` to v1.1.3, `vcf2cytosure` to v0.8 https://github.com/Clinical-Genomics/BALSAMIC/pull/987
+* toml config file for vcfanno https://github.com/Clinical-Genomics/BALSAMIC/pull/1012
+* Split `vep_germline` rule into `tumor` and `normal` https://github.com/Clinical-Genomics/BALSAMIC/pull/1018
+* Extract number of variants from clinical files https://github.com/Clinical-Genomics/BALSAMIC/pull/1022
+
+Fixed:
+^^^^^^
+* Reverted `pandas` version (from `1.3.5` to `1.1.5`) https://github.com/Clinical-Genomics/BALSAMIC/pull/1018
+* Mate in realigned bam file https://github.com/Clinical-Genomics/BALSAMIC/pull/1019
+* samtools command in merge bam and names in toml for vcfanno https://github.com/Clinical-Genomics/BALSAMIC/pull/1020
+* If statement in `vep_somatic_clinical_snv` rule https://github.com/Clinical-Genomics/BALSAMIC/pull/1022
+* Invalid flag second of pair validation error https://github.com/Clinical-Genomics/BALSAMIC/pull/1025
+* Invalid flag second of pair validation error using picardtools https://github.com/Clinical-Genomics/BALSAMIC/pull/1027
+* Samtools command for mergetype tumor https://github.com/Clinical-Genomics/BALSAMIC/pull/1030
+* `varcall_py3` container building https://github.com/Clinical-Genomics/BALSAMIC/pull/1036
+* Picard and fastp commands params and cluster config for umi workflow https://github.com/Clinical-Genomics/BALSAMIC/pull/1032
+* Set channels in `varcall_py3` container https://github.com/Clinical-Genomics/BALSAMIC/pull/1035
+* Delly command for tumor-normal analysis https://github.com/Clinical-Genomics/BALSAMIC/pull/1039
+* tabix command in bcftools_quality_filter_TNscope_umi_tumor_only rule https://github.com/Clinical-Genomics/BALSAMIC/pull/1040
+
+Removed:
+^^^^^^^^
+* case ID from the PON `.cnn` output file https://github.com/Clinical-Genomics/BALSAMIC/pull/983
+* `TNhaplotyper` for paired WGS analysis https://github.com/Clinical-Genomics/BALSAMIC/pull/988
+* `TNhaplotyper` for tumor only WGS analysis https://github.com/Clinical-Genomics/BALSAMIC/pull/1006
+* `TNhaplotyper` for TGS https://github.com/Clinical-Genomics/BALSAMIC/pull/1022
+
 [10.0.5]
 --------
 
@@ -9,9 +68,9 @@ Changed:
 
 [10.0.4]
 ---------
+
 Fixed:
 ^^^^^^
-
 * Increase `sambamba_exon_depth` rule run time https://github.com/Clinical-Genomics/BALSAMIC/pull/1001
 
 [10.0.3]